* net/loveruby/cflat/asm/Type.java: new class representing assembly integer types. * net/loveruby/cflat/asm/*.java: support it. * net/loveruby/cflat/ir/*.java: support it. * net/loveruby/cflat/compiler/IRGenerator.java: convert ast Type -> asm Type. * net/loveruby/cflat/compiler/CodeGenerator.java: use asm Type.
git-svn-id: file:///Users/aamine/c/gitwork/public/cbc/trunk@4160 1b9489fe-b721-0410-924e-b54b9192deb8

+++ net/loveruby/cflat/asm/Assembler.java
@@ -1,5 +1,4 @@
 package net.loveruby.cflat.asm;
-import net.loveruby.cflat.type.*;
 import net.loveruby.cflat.utils.*;
 import java.util.*;
 
@@ -92,7 +91,7 @@ public class Assembler {
     }
 
     protected String typeSuffix(Type t) {
-        switch ((int)t.size()) {
+        switch (t.size()) {
         case 1: return "b";
         case 2: return "w";
         case 4: return "l";
